The Prover interface has been changed to an abstract class.
The implementation of Prover#solve(String) is provided by prolog4j-api.

The place of arguments to be substituted to a goal can be denoted by "{}" signs following variable names in the goal string. (Prover#solve(String, Object...))

The implementation of the method is also provided by prolog4j-api, and it calls Prover#solve(String, String[], Object[]) which has been made protected. 

The Prolog4J implementations should only implement this solve method, but they should not make it public so that it does not interfere with Prover#solve(String, Object...).

Unit test adopted.